Round tab strip size to ints.
Bug: 1414229
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4235675
Commit-Queue: Sky Malice <skym@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1103372}

+++ chrome/browser/android/compositor/scene_layer/tab_strip_scene_layer.h
@@ -53,8 +53,8 @@ class TabStripSceneLayer : public SceneLayer {
 
   void UpdateTabStripLayer(JNIEnv* env,
                            const base::android::JavaParamRef<jobject>& jobj,
-                           jfloat width,
-                           jfloat height,
+                           jint width,
+                           jint height,
                            jfloat y_offset,
                            jboolean should_readd_background,
                            jint background_color);
@@ -63,8 +63,8 @@ class TabStripSceneLayer : public SceneLayer {
                         const base::android::JavaParamRef<jobject>& jobj,
                         jfloat x,
                         jfloat y,
-                        jfloat width,
-                        jfloat height,
+                        jint width,
+                        jint height,
                         jint color,
                         jfloat alpha);
